arm64: dts: imx91-var-dart: Update panel-timing
According to the VLCD-CAP-GLD-LVDS datasheet [1], the pixel clock
frequency should not exceed 40 MHz. Update the pixel clock frequency
from 71 MHz to 30 MHz, which is typical according to the datasheet
Set pixelclk-active to 1, as data is sampled on the falling edge
of the pixel clock and is driven on the rising edge of the pixel clock.

[1] https://www.variscite.com/wp-content/uploads/2017/12/VLCD-CAP-GLD-LVDS.pdf

Signed-off-by: Natalia Kovalenko <natasha.k@variscite.com>